+++ KmpIO/src/commonTest/kotlin/com/oldguy/common/io/FileTests.kt
@@ -207,6 +207,8 @@ class FileTests(testDirPath: String) {
             """.trimIndent() + eol
         val hexContent = Utf16LE().encode(textContent)
 
+        val macosIgnore = ".DS_Store"
+
         suspend fun testDirectory(): File {
             File.workingDirectory().up().apply {
                 assertTrue(exists)
@@ -251,7 +253,9 @@ class FileUnitTests {
         runTest {
             FileTests.testDirectory().apply {
                 println(fullPath)  // Use to find full path iosX64 and android tests are using.
-                directoryList().apply {
+                directoryList()
+                    .filter { it != FileTests.macosIgnore }
+                    .apply {
                     println(this)
                     assertEquals(8, size)
                     assertTrue { contains("ZerosZip64.zip") }
@@ -263,8 +267,10 @@ class FileUnitTests {
                     assertTrue { contains("dir2") }
                     assertTrue { contains("あ.png") }
                 }
-                directoryFiles().apply {
-                    assertEquals(7, size)
+                directoryFiles()
+                    .filter { it.name != FileTests.macosIgnore }
+                    .apply {
+                    assertEquals(8, size)
                     forEach {
                         assertTrue { it.exists }
                     }